requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
distro==1.9.0
elevenlabs==1.50.3
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.58.1
pydantic==2.10.4
pydantic_core==2.27.2
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
websockets==14.1
python-dotenv~=1.0.1
fastapi~=0.115.6
tabulate~=0.9.0
moviepy~=2.1.1
pytest~=8.3.4